requirements.txt



google.generativeai

langchain
langchain-google-genai
langchain-chroma
beautifulsoup4
requests
python-dotenv